use the index package instead of multind
This will necessitate a change in the way makeindex is run when building
a document that uses the cyber package. But the original way makeindex
was to be run looks never to have been documented properly, so the new
way is not either. In short, instead of running makeindex on a different
basename, you run makeindex jobname.cdx -o jobname.cnd, where jobname in
the context of CMITS is likely to be main. Judging from a look at shaney,
getting rid of multind will involve more changes to shaney than to cyber.